add more entry keys, basic color support, refactor entries
Entry text is now stored in an EntryText structure, present in all entries. previously each entry had its text stored in a char* in a structure pointed to by the data field in the Entry structure.

This made it a bit easier to add some light styling options like:
fgcolor, bgcolor, and colormode.

More details will be in the readme.